membership page backend done
