+++ src/app/membership/page.jsx
@@ -3,24 +3,8 @@
 import React, { useState } from "react";
 import { motion } from "framer-motion";
 import { FaCheckCircle } from "react-icons/fa";
-import { useSession } from "next-auth/react";
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  DialogFooter,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
-import {
-  Select,
-  SelectTrigger,
-  SelectContent,
-  SelectItem,
-  SelectValue,
-} from "@/components/ui/select";
+import MembershipModal from "./MembershipModal";
 
 const data = [
   {
@@ -88,46 +72,13 @@ const data = [
   },
 ];
 
-export default function Membership() {
-  const { data: session } = useSession();
-//   const session = { user: { name: "John Doe", email: "4KtJd@example.com", image: "https://example.com/avatar.jpg" } };
+export default function MembershipCards() {
   const [open, setOpen] = useState(false);
   const [selectedPlan, setSelectedPlan] = useState(null);
 
   const handleOpen = (plan) => {
     setSelectedPlan(plan);
     setOpen(true);
-  };
-
-  const handleFormSubmit = async(e) => {
-    e.preventDefault();
-    // Handle form submission logic here
-    const formData = new FormData(e.target);
-    const data = Object.fromEntries(formData);
-    data.plan = selectedPlan.name;
-
-    try {
-    const res = await fetch("/api/users", {
-      method: "PATCH",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(data),
-    });
-
-    const result = await res.json();
-    console.log(result);
-
-    if (result.success) {
-      alert("Membership data updated successfully!");
-    } else {
-      alert("Error updating data.");
-    }
-  } catch (err) {
-    console.error(err);
-    alert("Something went wrong.");
-  }
-    setOpen(false);
   };
 
   return (
@@ -142,7 +93,6 @@ export default function Membership() {
         </p>
       </div>
 
-      {/* Membership Cards */}
       <div className="grid gap-8 sm:grid-cols-1 md:grid-cols-2 lg:grid-cols-3">
         {data.map((plan) => (
           <motion.div
@@ -152,7 +102,6 @@ export default function Membership() {
             }`}
             whileHover={{ scale: 1.05 }}
           >
-            {/* Badge */}
             {plan.badge && (
               <span className="absolute top-4 right-4 bg-[#93EA33] text-white text-xs font-semibold px-3 py-1 rounded-full">
                 {plan.badge}
@@ -190,101 +139,14 @@ export default function Membership() {
         ))}
       </div>
 
-      {/* Modal Form */}
-      <Dialog open={open} onOpenChange={setOpen}>
-        <DialogContent className="min-w-2xl">
-          <DialogHeader>
-            <DialogTitle className={"text-2xl font-semibold"}>Join {selectedPlan?.name} Plan</DialogTitle>
-          </DialogHeader>
-
-          <form onSubmit={handleFormSubmit} className="grid gap-6 grid-cols-1 md:grid-cols-2">
-            {/* Name */}
-            <div>
-              <Label>Name</Label>
-              <Input name="name" value={session?.user?.name || ""} readOnly />
-            </div>
-
-            {/* Email */}
-            <div>
-              <Label>Email</Label>
-              <Input name="email" value={session?.user?.email || ""} readOnly />
-            </div>
-
-            {/* Height */}
-            <div>
-              <Label>Height (cm)</Label>
-              <Input name="height" type="number" placeholder="Enter height" />
-            </div>
-
-            {/* Current Weight */}
-            <div>
-              <Label>Current Weight (kg)</Label>
-              <Input name="currentWeight" type="number" placeholder="Enter current weight" />
-            </div>
-
-            {/* Current Fat % */}
-            <div>
-              <Label>Current Fat %</Label>
-              <Input name="currentFatPercentage" type="number" placeholder="Enter current fat percentage" />
-            </div>
-
-            {/* Current Muscle Mass */}
-            <div>
-              <Label>Current Muscle Mass (kg)</Label>
-              <Input name="currentMuscleMass" type="number" placeholder="Enter current muscle mass" />
-            </div>
-
-            {/* Goal */}
-            <div>
-              <Label>Goal</Label>
-              <Select name="goal">
-                <SelectTrigger>
-                  <SelectValue placeholder="Select goal" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="bulking">Bulking</SelectItem>
-                  <SelectItem value="cutting">Cutting</SelectItem>
-                  <SelectItem value="weightLoss">Weight Loss</SelectItem>
-                  <SelectItem value="muscleGain">Muscle Gain</SelectItem>
-                  <SelectItem value="endurance">Endurance</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-
-            {/* Goal Weight */}
-            <div>
-              <Label>Goal Weight (kg)</Label>
-              <Input name="goalWeight" type="number" placeholder="Enter goal weight" />
-            </div>
-
-            {/* Goal Fat % */}
-            <div>
-              <Label>Goal Fat %</Label>
-              <Input name="goalFatPercentage" type="number" placeholder="Enter goal fat percentage" />
-            </div>
-
-            {/* Goal Muscle Mass */}
-            <div>
-              <Label>Goal Muscle Mass (kg)</Label>
-              <Input name="goalMuscleMass" type="number" placeholder="Enter goal muscle mass" />
-            </div>
-
-            {/* Extra field for Wellness Plus */}
-            {selectedPlan?.name === "Wellness Plus" && (
-              <div>
-                <Label>Your Disabilities / Health Issues</Label>
-                <Input name="disabilities" placeholder="Describe your conditions" />
-              </div>
-            )}
-
-            <DialogFooter className={"col-span-1 md:col-span-2 mx-auto"}>
-              <Button type="submit" className="bg-[#93EA33] hover:bg-[#77b036]">
-                Submit
-              </Button>
-            </DialogFooter>
-          </form>
-        </DialogContent>
-      </Dialog>
+      {/* Modal */}
+      {selectedPlan && (
+        <MembershipModal
+          open={open}
+          setOpen={setOpen}
+          selectedPlan={selectedPlan}
+        />
+      )}
     </section>
   );
 }
